# website/data/openapi/dotmarc-api.json
{
  "components": {
    "schemas": {
      "ApiAcknowledgement": {
        "example": {
          "ticketClosed": true,
          "ticketsClosed": 1,
          "ticketsFailed": 0
        },
        "properties": {
          "ticketClosed": {
            "type": "boolean"
          },
          "ticketsClosed": {
            "format": "int32",
            "type": "integer"
          },
          "ticketsFailed": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "ticketClosed",
          "ticketsClosed",
          "ticketsFailed"
        ],
        "type": "object"
      },
...
      "ApiAlert": {
        "example": {
          "acknowledgeable": false,
          "domain": {
            "id": 42,
            "name": "contoso.example"
          },
          "id": 901,
          "message": "contoso.example's SPF record has more than 10 DNS lookups.",
          "raisedUtc": "2026-10-01T09:30:00+00:00",
          "resolved": false,
          "resolvedUtc": null,
          "severity": "Warning",
          "subject": "contoso.example",
          "title": "SPF record broken",
          "type": "SpfRecordBroken",
          "typeName": "SPF record broken"
        },
        "properties": {
          "acknowledgeable": {
            "type": "boolean"
          },
          "domain": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ApiNamedRef"
              }
            ],
            "type": "object"
          },
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "raisedUtc": {
            "format": "date-time",
            "type": "string"
          },
          "resolved": {
            "type": "boolean"
          },
          "resolvedUtc": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "typeName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "typeName",
          "subject",
          "domain",
          "severity",
          "title",
          "message",
          "raisedUtc",
          "resolved",
          "resolvedUtc",
          "acknowledgeable"
        ],
        "type": "object"
      },
...
      "ApiNamedRef": {
        "properties": {
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "type": "object"
      },
      "ApiPageOfApiAlert": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApiAlert"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "pageSize": {
            "format": "int32",
            "type": "integer"
          },
          "totalCount": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object"
      },
...
      "HttpValidationProblemDetails": {
        "properties": {
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "errors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "ProblemDetails": {
        "properties": {
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "dmk_ API key",
        "description": "An API key created on dotMARC's Access page.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Read domains, DNS health, DMARC report summaries, groups, tags and alerts, and add, import and organise domains. Authenticate with an API key from dotMARC's Access page: `Authorization: Bearer dmk_...`. Each operation needs the permission named in its description, and a key limited to certain groups only sees their domains. Each key may make 120 requests a minute.",
    "title": "dotMARC API",
    "version": "0.9.0"
  },
  "openapi": "3.0.4",
  "paths": {
    "/api/v1/alerts": {
      "get": {
        "description": "Alerts newest first, open ones by default (status=all includes resolved), 50 a page by default and at most 200. A key limited to certain groups sees only alerts about its domains.\n\nNeeds the AlertsView permission.",
        "operationId": "ListAlerts",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiPageOfApiAlert"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            },
            "description": "The request couldn't be used. `errors` names each field that's wrong and why."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "No API key was sent, or the key is unknown, expired or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The key's role doesn't have the permission this operation needs, or the key is limited to certain groups and this would reach outside them. `detail` says which."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "This key has made more requests this minute than dotMARC allows (120 by default). Wait for the number of seconds in the Retry-After header, then try again."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Something went wrong in dotMARC that the request didn't cause. It's safe to retry a read; check before retrying a change. If it keeps happening, dotMARC's logs have the details."
          }
        },
        "summary": "List alerts",
        "tags": [
          "Alerts"
        ],
        "x-dotmarc-permission": "AlertsView"
      }
    },
    "/api/v1/alerts/{id}/acknowledge": {
      "post": {
        "description": "Closes a DMARC policy weakened or nameservers changed alert and accepts the current value as normal, closing its PSA ticket too. Other alerts close themselves once fixed, so acknowledging them is a conflict. ticketClosed is false when the ticket couldn't be closed and needs closing by hand.\n\nNeeds the AlertsManage permission.",
        "operationId": "AcknowledgeAlert",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAcknowledgement"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "No API key was sent, or the key is unknown, expired or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The key's role doesn't have the permission this operation needs, or the key is limited to certain groups and this would reach outside them. `detail` says which."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "There's no such item, or this key can't see it. A key limited to certain groups can't see what's outside them."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request conflicts with how things are now, such as a domain that's already in dotMARC or an alert that can't be acknowledged. `detail` says what."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "This key has made more requests this minute than dotMARC allows (120 by default). Wait for the number of seconds in the Retry-After header, then try again."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Something went wrong in dotMARC that the request didn't cause. It's safe to retry a read; check before retrying a change. If it keeps happening, dotMARC's logs have the details."
          }
        },
        "summary": "Acknowledge an alert",
        "tags": [
          "Alerts"
        ],
        "x-dotmarc-permission": "AlertsManage"
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}